.tap that path
The `lib/install/install.rb` script is missing `.tap` in the append to `config/initializers/assets.rb` file and so does actually set any of the passed assets/directories. The tests did not pick up on this, because the dummy app correctly uses .tap in it.

Thanks again for blazing the trail on ESM and Rails! Love that it's been extracted to a separate repo :)