Make rescuable asset error classes configurable
Consumers may be using a different asset pipeline than sprockets-rails,
which means that when `Importmap::Map#resolve_asset_paths` invokes the
application's `asset_path` helper method, it may need to rescue
different exceptions than `Sprockets::Rails::Helper::AssetNotFound`. To
support that, we've added a `rescuable_asset_errors` property to the
importmap configuration options. We add
`Sprockets::Rails::Helper::AssetNotFound` to the array if
`Sprockets::Rails` is defined, which should allow the gem to work the
same as before in applications that use sprockets-rails. But now it'll
also be usable in applications that use something other than
sprockets-rails.