Update the regex
The failed tests are now passing.

The approach is a little naive in that a string surrounded by one double quote and one single quote will return as a positive result.

Whether this is important enough that it should be addressed with a more complex regex is up for question.